requirements.txt
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.3
Pillow>=9.4.0
PyQt5==5.15.9
PyQt5_sip==12.12.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0